backend/requirements.txt
# WhispyrKeep Backend Dependencies
# Django + DRF Stack

# Core Framework
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0

# Database
psycopg[binary]>=3.1,<4.0
dj-database-url>=2.1,<3.0

# Async Tasks
celery>=5.3,<6.0
redis>=5.0,<6.0

# Vector Database
chromadb>=0.4,<1.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0

# Security
cryptography>=42.0,<43.0

# Utilities
python-dotenv>=1.0,<2.0

# Development
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0

# Production
gunicorn>=21.0,<23.0
